[bouqueau] constants.h should not include any other GPAC headers
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@3651 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ modules/amr_dec/amr_in.c
@@ -464,17 +464,17 @@ fetch_next:
 		switch (read->mtype) {
 		case TYPE_AMR:
 			ft = (toc >> 3) & 0x0F;
-			read->data_size = GF_AMR_FRAME_SIZE[ft];
+			read->data_size = (u32)GF_AMR_FRAME_SIZE[ft];
 			break;
 		case TYPE_AMR_WB:
 			ft = (toc >> 3) & 0x0F;
-			read->data_size = GF_AMR_WB_FRAME_SIZE[ft];
+			read->data_size = (u32)GF_AMR_WB_FRAME_SIZE[ft];
 			break;
 		default:
 			for (i=0; i<GF_SMV_EVRC_RATE_TO_SIZE_NB; i++) {
 				if (GF_SMV_EVRC_RATE_TO_SIZE[2*i]==toc) {
 					/*remove rate_type byte*/
-					read->data_size = GF_SMV_EVRC_RATE_TO_SIZE[2*i+1] - 1;
+					read->data_size = (u32)GF_SMV_EVRC_RATE_TO_SIZE[2*i+1] - 1;
 					break;
 				}
 			}
